--- assets/index.css
@@ -1,75 +1,0 @@
-*{
-  box-sizing: border-box;
-  word-break: keep-all;
-  font-family: "Arial", serif;
-}
-
-body,main{
-  display: flex;
-  flex-direction: column;
-  align-content: flex-start;
-}
-
-body{
-  margin: auto;
-  height: auto;
-  min-height: 100vh;
-  min-width: 130px;
-  background: whitesmoke;
-  max-width: 1200px;
-}
-
-#content{
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  justify-content: center;
-  align-content: center;
-  background-size:60px 60px;
-  height: auto;
-  min-height: 99vh;
-}
-
-.addNewArticleButton{
-  cursor: pointer;
-  position: fixed;
-  bottom: 5px;
-  left: 5px;
-  height: 30px;
-  width: 30px;
-  border-radius: 15px;
-  padding-left: 6px;
-  font-size: 30px;
-  color: white;
-  background-color: black;
-  z-index: 4;
-}
-
-input{
-  border: none;
-  background-color: inherit;
-  margin-bottom: 5px;
-  color: gray;
-}
-
-@media (min-width: 600px) {
-
-  #content,main{
-    flex-direction: row;
-  }
-
-  main{
-    width: 100%;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    padding-left: 2.3%;
-    padding-top: 2%;
-  }
-
-  .addNewArticleButton{
-    font-weight: bold;
-    left: auto;
-    right: 0;
-    margin: 2% 20px 0 1%;
-  }
-}
