+++ assets/index.html
@@ -1,6 +1,19 @@
 <!DOCTYPE html >
-<html lang="en" style = "background: whitesmoke;">
+<html lang="en" >
 <head>
+  <style>
+    body, html, #root{
+      height: auto;
+      min-height: 100vh;
+      width: 100%;
+      overflow: hidden;
+      background: whitesmoke;
+      margin: auto;
+      box-sizing: border-box;
+      word-break: keep-all;
+      font-family: "Arial", serif;
+    }
+  </style>
   <meta charset="utf-8" />
   <link rel="shortcut icon" href="icon.png" />
   <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no"/>
@@ -10,7 +23,7 @@
   <title>(Reacе-Redux)I do it</title>
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css">
 </head>
-<body onoffline="" style="margin: auto">
+<body onoffline="">
 <div id="root" >
 </div>
 </body>
