+++ src/api/dashboard.js
@@ -12,6 +12,5 @@ export const addDashboard = newDashboard => axios.post('/api/todolists', newDash
 
 export const updateList = (id, newList) => axios.put(`/api/todolists/${id}`, newList);
 
-export const shareTodoListToUser = (id, userName) =>
-    axios.post(`/api/todolists/${id}/share?username=${userName}`);
-
+export const shareTodoListToUser = (id, userName) => (
+    axios.post(`/api/todolists/${id}/share?username=${userName}`));
