+++ src/index.js
@@ -1,12 +1,12 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import 'regenerator-runtime/runtime';
 import React from 'react';
-import { Provider } from 'react-redux';
 import ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
 import { PersistGate } from 'redux-persist/integration/react';
 import { Router } from 'react-router-dom';
-import './api/dashboard';
 import App from './App';
+import './api/dashboard';
 import history from './config/history';
 import { store, persistor } from './config/store';
