+++ src/scenes/header/Component.styles.js
@@ -3,38 +3,31 @@ import styled from 'styled-components';
 export const App = styled.div`
         display: flex;
         flex-direction: column;
-        align-content: center;
+        flex: auto;
+        height: 100vh;
+        overflow auto;
 `;
 
 export const Header = styled.header`
-        display: flex;
-        height: 60px;
-        width: auto;
-        flex-direction: row;
-        padding:30px 0;
-        margin-bottom: 20px;
-
+    display: flex;
+    flex-direction: row;
+    padding:10px;
+    box-shadow: 0 0  15px 0  rgba(0,0,0,0.2);
+    z-index:2;
+    background-color: white;
     & > b{
         font-size: 42px;
     }
-    & > div{
-        margin-top: 0.8%;
-    }
+}
 `;
 
 export const Container = styled.div`
-   margin: auto;
-   box-sizing: border-box;
-   word-break: keep-all;
-   font-family: "Arial", serif;
-   height: auto;
-   min-height: 100vh;
+   display: flex;
+   flex-direction: column;
    min-width: 130px;
-   background: whitesmoke;
-   max-width: 1200px;
    input{
          border: none;
          background-color: inherit;
          color: grey;
-    }
+   }
 `;
